+++ src/printer_mlwtree.ml
@@ -141,7 +141,7 @@ let pp_type fmt typ =
       | Tycoll i      -> (String.capitalize_ascii i) ^ ".collection"
       | Tyview i      -> (String.capitalize_ascii i) ^ ".view"
       | Typartition i -> (String.capitalize_ascii i) ^ ".view"
-      | Tysubset i    -> (String.capitalize_ascii i) ^ ".view"
+      | Tyaggregate i -> (String.capitalize_ascii i) ^ ".view"
       | Tymap i       -> "map " ^ i
       | Tyrecord i    -> i
       | Tyasset i     -> i